+++ NEWS.md
@@ -8,6 +8,9 @@
 - New `label_placement_minmax()` to label contours at the maximum and minimum 
 points of a contour (mimicking [isoband's behaviour](https://wilkelab.org/isoband/articles/isoband3.html))
 
+- `geom_tanaka()` now has a (a bit experimental) argument `smooth` which allows to smooth
+the transition between segments. 
+
 
 ## Bugfixes